Fixes the Donut Box Icon for Mapping (#65908)
Hey there,

Look over here:

This is so awfully undescript that anyone who doesn't map doesn't "get" that this is supposed to be a donut box. It also doesn't help if you're trying to position stuff on a table around a donut box, and you want stuff to look pretty without having to compensate for an unknown variable (only known when you compile).

So, I made a composite icon of both the inner and outer shells of the box, and set that for the icon_state in the donut box's code. It uses base_icon_state to generate the donuts being filled into the box, and when I tested it on local, everything still worked as intended.