Access Helpers Pt. 3/6(?) - Kilo Station (#66805)
This PR continues to replace door accesses with mapping helpers with the same design standards as #66588 and #66753 but this time with KiloStation

Some removed inconsistencies:
-Random centcomm codes on maintenance doors leading from the centcomm docking port
-Chaplain access on the sec cremator
-Lawyer access to the gear room lobby, medbay, evidence, interrogation, and gulag areas
-Chaplain access on morgue doors (If we want this they should just have morgue access)
-Research and Mining on the aux base
-Research Director on the tech room
-Heads on the secure tech room
-Slew of accesses on service maints instead of general service access

Standardizes access requirements for the map and reduces changes for inconsistent mapping to happen by use of mapping helpers. Makes access requirements easier to audit.