Makes Cultist Summoning Nar'Sie An Important Announcement (#67143)
Hey there,

I first noticed this about three months, but I was pretty pleased about it since I was a cultist. I decided that today I should

Fix #Issue

It just makes it an important message, so that it will indeed reveal the location that the cult is summoned, and won't be muffled by that good-for-nothing potato-chip-eating centcom intern.