Fixes hair dye spray not updating your sprite after being applied (#67237)
The proc that updates your sprite after dying your hair using the hair dye spray was missing an argument. This meant that while the new dye would be applied, it wouldn't be reflected in your sprite until something else caused your head sprite to refresh, such as changing your hairstyle on a mirror.