Restores the icebox sci break room to a decent-er standard (#67344)
So back when the ordnance room of icebox got remapped, they smudged around the breakroom I added a few months ago.

Unfortunately, they left it in, well, quite a state.

There's an unconnected scrubber underneath the couch, and all the smooth-large tiles that used to designate where the scrubbers and vents were to be placed weren't moved at all! The room was left looking real sad and empty.