Testing with an experimental version of auxlua
This version of auxlua is built using a branch of my fork of auxtools in which the dll is not pinned, and can thus be freed. This change, along with some others that ensure that unloading auxlua won't crash DD, will allow for `auxlua.dll` to play nice with tgs's a/b switching.