USB cables can now be attached directly to bluespace launchpads (#68616)
See title. Because of this change, a launchpad does not need a connected console if you have no need for it. You can still use the console to control a launchpad, and you can still connect a USB cable to a console to control all its connected launchpads with a single circuit component.