[MIRROR] [MDB IGNORE] Bananas now have a 1% chance of spawning with the boomerang component (#2119)
* Bananas now have a 1% chance of spawning with the boomerang component (#69325)

Bananas now have a 1% chance to be bananarangs on initialization. The description is altered to make it slightly easier to notice, but otherwise there is no indication beyond throwing it.

bananarang hehe :)

Can be used for entertainment, bargaining, or engaging with a rarely seen component within the sandbox.


Bananas now have a 1% chance to be boomerang-bananas. These look identical to normal bananas, aside from a slightly unique examine text. Make sure to check your bananas before peeling it to slip the HOS, you may just have a rare item on your hands.

* Bananas now have a 1% chance of spawning with the boomerang component

Co-authored-by: Rhials <Datguy33456@gmail.com>