[MIRROR] [MDB IGNORE] Open turfs no longer care about hotspots when they reach melting point. (#2197)
* Open turfs no longer care about hotspots when they reach melting point. (#69269)

Removes /turf/open/burn proc. This removal will cause the parent proc to get called when it gets called on an open turf, without the hotspot check.

Lemon notes:
This does actually change a lot, but I want it anyway
It was initially like this because only hotspots could be a source of melting. when I changed that with my linda refactor, I kept the no hotspot requirement because it A: was close to the same behavior, and B: in theory meant less excited group rebuilds.

I have since decided that rebuilds like that especially in the context of hotspots are ok.

* Open turfs no longer care about hotspots when they reach melting point.

Co-authored-by: Pickle-Coding <58013024+Pickle-Coding@users.noreply.github.com>