[MIRROR] [MDB IGNORE] adds logging for the Aurora Caelus oven ignition (#2235)
* adds logging for the Aurora Caelus oven ignition (#69525)

About The Pull Request

So apparently during the Aurora Caelus event there's a 1% chance that the oven in the kitchen ignites. This adds logging for when that does end up happening.

Why It's Good For The Game

Not having this logged can make it really hard to investigate why the entire kitchen burned down. I only found out that this caused the fire because I happened to be in view and saw an "oh egads" balloon alert for a split second. I almost thought I imagined it or something and it turns out it's a simpsons joke...

* adds logging for the Aurora Caelus oven ignition

Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>