[MIRROR] [MDB IGNORE] visualizes *cry emote + getting peppersprayed makes you cry (#2271)
* visualizes *cry emote + getting peppersprayed makes you cry (#69492)

About The Pull Request

Using the *cry emote makes tears come out of your eyes for the next 12.8 seconds (4 loops of the animation). You now cry when you get sprayed with pepperspray. See the demo here.
image

The crying overlay is placed by the eyes organ. I went back to the blush overlay and made it so that it's placed by the head bodypart. Also the blush overlay now gets adjusted based on OFFSET_FACE.

Why It's Good For The Game

It makes the characters more visually expressive. The crying gives pepperspray some more visual feedback, not just when you get hit but while you're recovering from it afterwards as well.

* visualizes *cry emote + getting peppersprayed makes you cry

Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>