[MIRROR] [MDB IGNORE] Heads are no longer tasked to kill their own pets (#2283)
* Heads are no longer tasked to kill their own pets (#69262)


About The Pull Request

Someone on my server complained that they got assigned as CE to kill Poly, and I have taken a look, and printed the contents of the possible_heads list, and turns out, your own will never gets removed from it. This PR fixes that, by removing your own role from the off chance it is contained in the list of possible targets.

I consider this a fix, but if it is a balance issue, I will update the PR accordingly. It is balance now.
Why It's Good For The Game

Heads killing their own pets is an extremely low effort objective. Not to mention, they are their beloved companion...
Changelog

cl
balance: Heads are no longer tasked to kill their own pets
/cl

* Heads are no longer tasked to kill their own pets

Co-authored-by: Profakos <profakos@gmail.com>