[MIRROR] [MDB IGNORE] Add air alarm and atmos machinery to kilo server room + meta tcommsat air alarm pixel shift (#2336)
* Add air alarm and atmos machinery to kilo server room + meta tcommsat air alarm pixel shift (#69589)

Replaces the intercom in the Kilo Research Department Server Room with an Air Alarm and adds a vent and scrubber. Also removes the air alarm from the server/cold side of the room which was going off constantly and making atmos techs mad.

Also, I shifted the tcommsat air alarm on metastation 2 pixels to the left because it didn't match the rest of the air alarms. Yes, really.

* Add air alarm and atmos machinery to kilo server room + meta tcommsat air alarm pixel shift

Co-authored-by: msgerbs <msgerbs@users.noreply.github.com>