[MIRROR] [MDB IGNORE] Bumps auxlua to 1.3.0 (#2341)
* Bumps auxlua to 1.3.0 (#69681)

Bumps auxlua to 1.3.0. The only thing this adds is the loadstring function, which is present in basically all other forms of lua scripting.
loadstring allows for compiling strings of lua code and running them. It's a function that exists in most other forms of lua. One particular use case is downloading the raw code of a lua module from the internet using a /datum/http_request, then using loadstring to run that module.

* Bumps auxlua to 1.3.0

Co-authored-by: Y0SH1M4S73R <legoboyo@earthlink.net>