Adds randomly generated mime names (#69791)
About The Pull Request

Adds 26 new names for mimes that can be found by randomizing your name in the character editor.

Why It's Good For The Game

There's currently 5 times as many random clown names as there are random mime names (124 to 24, specifically). This PR helps close the gap, which adds variety to the names of mime players, whether they're hardcore random or their player just couldn't think of a name.