Updates Liver Failure Eyes Examine Text (#69967)
So when livers fail it adds something about eyes to the carbon's examine text. The 2nd state of eye yellowness used a static he, that has been replaced by a helper to get the correct pronoun. The 3rd state didn't capitalise the first pronoun use and followed it up with a static they, that has also been replaced by a pronoun helper.