[MIRROR] [MDB IGNORE] Removes Double "Random Event Triggering" message (#2692)
* Removes Double "Random Event Triggering" message (#70137)

* Removes Double "Random Event Triggering" message

Hey there,

This has been bugging me for a solid while now, so I decided to fix it today. Apparently, we had the same exact thing trigger at about the same exact time since we were running this check (if similar criteria were fulfilled) at both the pre-event setup and the post-event setup. Both procs were even sleeping as well, so I thought this was all a bit silly and just rewrote the code on the pre-event setup while still retaining the functionality on the actual event setup (since that will sleep anyways).

* Adds Balloon Alerts To Valve Open/Closing

Hey there,

I was talking with someone trying to get the hang of atmospherics, and they were a bit confused on why their digital valve didn't seem to be "on". Well, the thing is that it was on, they just didn't know to peek at that small little green dial on the digital valve. Although the animations are quite explanatory to someone like me who's been doing this for 3+ years, I definitely can see how it's more confusing to someone new. So, let's toss in a small balloon alert (that only plays after the animation is complete, as to not interfere with the nice spritework) to provide a nice alert as to what exactly they're doing.

* Revert "Adds Balloon Alerts To Valve Open/Closing"

This reverts commit aee7ab44d4e0ba9bb57f79724d18fbb60c1a6b04.

* why do we even check it that way

* reworks it to retain babyproofing functionality

i spoke it through with some other people and I think this is probably the best way to do it

* wait that was wrong

* re-adds alert_observers check

* inverts the behavior to be on PreRunEvent

mkay

* Removes Double "Random Event Triggering" message

Co-authored-by: san7890 <the@san7890.com>