[MIRROR] [MDB IGNORE] Tramstation/Icebox now have EOD closets where they should be (#2769)
* Tramstation/Icebox now have EOD closets where they should be (#70322)

Tramstation and Icebox toxins now have EOD closets in their vicinity. While working on this, I noticed that Tramstation lacked a security EOD closet as well, and fixed that. Nothing got moved around except for a potted plant and some vents.

A report has already been submitted to Space OSHA regarding the lack of appropriate PPE in the workplace. Space Station 13 is supposed to be a SAFE WORKING ENVIRONMENT.

* Tramstation/Icebox now have EOD closets where they should be

Co-authored-by: Rhials <Datguy33456@gmail.com>