[MIRROR] [MDB IGNORE] Save 0.1s in generate_station_area_list (#3069)
* Save 0.1s in generate_station_area_list (#70933)

`area.contents` loops over everything in the world to collect its list,
every time.

We need to avoid it where possible. Lemon already working on a PR to
cache it. In the distant future we will lint this.

This was a useless consumer of it. It cost 0.1s to index contents twice
like this.

* Save 0.1s in generate_station_area_list

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>
Co-authored-by: Jolly <70232195+Jolly-66@users.noreply.github.com>