[MIRROR] [MDB IGNORE] Adds REAGENT_CAN_BE_SYNTHESIZED to determination, universal indicator, and organic slurry (#3083)
* Adds REAGENT_CAN_BE_SYNTHESIZED to determination, universal indicator, and organic slurry (#70867)

## About The Pull Request

The determination, universal indicator, and organic slurry reagents now
have the REAGENT_CAN_BE_SYNTHESIZED flag, like most other reagents in
the game. This allows those reagents to be harvested from corpses (if
the corpse had them in their system when they died), to appear in
randomized chemical plant traits, and to be replicated by Odysseuses
(Odyssi?).

## Why It's Good For The Game

To my understanding, every reagent in the game is supposed to have the
REAGENT_CAN_BE_SYNTHESIZED flag unless there's a good reason to
specifically exclude them from being harvested or replicated.


![image](https://user-images.githubusercontent.com/42606352/198740606-ec0f267e-a431-49f4-aa9c-5a0fec5c5911.png)

Determination is a chemical that's generated when you receive a wound.
Its effects are incredibly minor and mostly wound-related (you can see
them here:
https://github.com/tgstation/tgstation/blob/master/code/modules/reagents/chemistry/reagents/other_reagents.dm#L2604).
Pretty much the only reason to go through the lengthy process of
harvesting determination (which involves torturing someone until they
generate it, then killing them, grinding them up, and extracting the
chem from their meat) is for the coolness factor, and I think that
should be respected. There are also no balance concerns with making
determination available to players, it's entirely outclassed by
roundstart chems, let alone chems that require a similar amount of
effort to obtain (such as bath salts).

Universal indicator is a chemical that changes color based on the pH of
the chems you mix it with. I think its inability to be synthesized is
genuinely just an oversight.

Organic slurry (aka yuck) is a chemical that's generated by dipping a
dead rat in welding fuel. It's that easy. It really shouldn't be so
complex that an Odysseus can't synthesize it.

## Changelog
:cl: ATHATH
qol: The determination, universal indicator, and organic slurry reagents
now have the REAGENT_CAN_BE_SYNTHESIZED flag, like most other reagents
in the game. This allows those reagents to be harvested from corpses (if
the corpse had them in their system when they died), to appear in
randomized chemical plant traits, and to be replicated by Odysseuses
(Odyssi?).
/:cl:

* Adds REAGENT_CAN_BE_SYNTHESIZED to determination, universal indicator, and organic slurry

Co-authored-by: ATH1909 <42606352+ATH1909@users.noreply.github.com>
Co-authored-by: Zephyr <12817816+ZephyrTFA@users.noreply.github.com>