[MIRROR] [MDB IGNORE] Allows mobs caught in disposals to escape once they have stopped moving - Take two (#3108)
* Allows mobs caught in disposals to escape once they have stopped moving - Take two (#70908)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

This is a re-opening of #69960, after addressing all prior reviews
(except one, see below).

## About The Pull Request
- A mob inside a disposalpipe holder object that has stopped can use the
"resist" action (hotkey: B by default) to struggle out. This will take
20 seconds (on top of the time spent waiting for their holder to stop),
and will play a sound. The pipe they are in will burst (taking the tile
with it if applicable), and deal a random 5 to 15 brute damage to all
mobs in the holder.
- When merging a moving holder with a stopped holder, pipes will now
check if the stopped holder has a mob, and if so, will merge the moving
holder into the stopped holder rather than the reverse. Thus, all new
mobs will become stuck on a stopped mob.
- When deleted, holders will now drop their contents to the floor first.
This is to prevent a very rare bug deleting all holder contents
(including mobs) if the holder attempts to move through a pipe currently
being destroyed.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Disposal traps are fun, but there shouldn't be a scenario where a player
is trapped indefinitely, with no ability to use tools or other items to
escape, that can be made for free using round-start items. If you want
to make a trap, the disposals pipes should be the delivery system, not
the destination.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: You can now resist out of disposals pipes once you have stopped, at
the cost of some brute damage.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* Allows mobs caught in disposals to escape once they have stopped moving - Take two

Co-authored-by: zxaber <37497534+zxaber@users.noreply.github.com>