[MIRROR] [MDB IGNORE] Add: Deaf Personnel Pin (#3141)
* Add: Deaf Personnel Pin (#71073)

This PR contains a single small addition: The deaf personnel pin/accessory for users of the Deaf quirk. I changed the quirk type to `item_quirk` in order to easily grant the pin to the user. Upon spawning, the quirk will put a deaf personnel pin in their backpack or hands, allowing them to pin it to their uniform as they please.

* Add: Deaf Personnel Pin

Co-authored-by: Dani Glore <fantasticdragons@gmail.com>