[MIRROR] [MDB IGNORE] Fixes caravan ambush ruin's id (#3146)
* Fixes caravan ambush ruin's id (#71105)

During initialization, when the Caravan Ambush space ruin loaded, there
have been error message during initialization, such as
`create_network_simple: SPACE/CARAVANAMBUSH.DOORS.AIRLOCKS IS INVALID,
replacing with LIMBO`. This was caused by the `/` present in the space
ruin's ID, which the network validation regex did not accept. I removed
`space/` from the beginning, both to make it compliant with the regex,
and also because there is no other kind of caravan ambushes to
differentiate it from.

* Fixes caravan ambush ruin's id

Co-authored-by: Profakos <profakos@gmail.com>