[MIRROR] [MDB IGNORE] samus bombs deal the same base damage to objects and mobs (#3246)
* samus bombs deal the same base damage to objects and mobs (#71215)

## About The Pull Request

reduces the damage modsuit bombs do vs. objects from 15 damage to 12.

## Why It's Good For The Game

whoever made modsuit bombs decided to define a damage variable with an
autodoc comment of "Damage done on explosion.", then _only use it for
object damage_, with a *hardcoded value of 12* being used for the damage
to mobs. it's a 3 point difference in damage, so I figured that the
better solution would just be to just unify the damage values instead of
making a new variable for this.

an alternative I considered was buffing the damage vs. mobs/hostile mobs
from 12/48 to 15/60, but idk if that'll affect lavaland balance in a bad
way. if it won't, go ahead and tell me and I'll do that.

## Changelog

:cl: ATHATH
balance: The damage modsuit bombs do to objects has been reduced from 15
damage to 12, to match the damage they deal to non-hostile mobs.
/:cl:

* samus bombs deal the same base damage to objects and mobs

Co-authored-by: ATH1909 <42606352+ATH1909@users.noreply.github.com>