[MIRROR] [MDB IGNORE] Fix crash of TGUI fax panel when there is no fax in the network. (#3248)
* Fix crash of TGUI fax panel when there is no fax in the network. (#71262)

## About The Pull Request

Thanks to @Bobbanz1 for letting me know about this error.
When designing faxes, I didn't foresee a situation where the fax list
could be empty. This caused the TGUI panel to crash. In this PR I fixed
this situation by checking for emptiness.

![FeUbGclCUq](https://user-images.githubusercontent.com/88540658/201709838-bb07c72f-4ee8-4db7-b17a-d9f82aae96a0.png)

## Changelog
:cl:
fix: Fix crash of TGUI fax panel when there is no fax in the network
/:cl:

Co-authored-by: twilightwanderer <twilightwanderer@users.noreply.github.com>

* Fix crash of TGUI fax panel when there is no fax in the network.

Co-authored-by: twilightwanderer <88540658+twilightwanderer@users.noreply.github.com>
Co-authored-by: twilightwanderer <twilightwanderer@users.noreply.github.com>