[MIRROR] [MDB IGNORE] Adding the ability to load the configuration of custom maps. (#3250)
* Adding the ability to load the configuration of custom maps. (#71277)

## About The Pull Request

By default, all custom maps are assigned one standard configuration.
This option rather limits the possibilities for creating custom levels.
For example, you can not use multilevel maps, because with this
configuration will always be loaded only one level. With this PR I want
to provide the ability to load a configuration file for a custom map
similar to the standard map.json configs.

![ZpdpfZWq2V](https://user-images.githubusercontent.com/88540658/201946344-fdf8589c-0e20-475c-a9d8-b21c9ff32dd8.png)

## Why It's Good For The Game

Allows administrators to flexibly configure custom maps. For example for
events, etc.

## Changelog
:cl:
admin: Adding the ability to load the configuration of custom maps
/:cl:

Co-authored-by: twilightwanderer <twilightwanderer@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>

* Adding the ability to load the configuration of custom maps.

Co-authored-by: twilightwanderer <88540658+twilightwanderer@users.noreply.github.com>
Co-authored-by: twilightwanderer <twilightwanderer@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>