[MIRROR] [MDB IGNORE] Updates ice cabin kitchen with modern equipment. (#3333)
* Updates ice cabin kitchen with modern equipment. (#71088)

## About The Pull Request

At the moment the ice cabin is quite outdated in the way that it only
has microwaves as methods to cook, this is extra awkward as most of the
food you get in this away method is meat. This updates the room by
including a griddle and oven to make some nice edible food.


![image](https://user-images.githubusercontent.com/16896032/200107098-0172ca42-44c1-4c65-abce-8b9fe103427d.png)


## Why It's Good For The Game

Do you want to spend a round in an ice cabin where the only food source
is eggs boiled in a microwave?

## Changelog


:cl:
fix: Updates the Ice cabin kitchen to be more modern. An oven and
griddle have been added.
/:cl:

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>

* Updates ice cabin kitchen with modern equipment.

Co-authored-by: carshalash <carshalash@gmail.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>