Janitor can use wagon key to stab people (#71696)
## About The Pull Request
The key can now be used to attack people by stabbing or throwing. I
always thought it was weird how most other roles have starter weapons.
Well the key's not a good weapon it's very weak but hey at least it's
something.


## Why It's Good For The Game
Little details and fluff is always nice, being able to stab people with
keys is pretty realistic, it's even used in self defense booklets! In
general it is a real thing and weird that we couldn't do it in the first
place. Being able to key people adds more realistic interaction that is
not overpowered.


## Changelog



:cl:
add:  Key can now hurt people by stabbing or throwing.
code: changed the keys code to make it able to hurt ppl
/:cl: