[MIRROR] [MDB IGNORE] Don't update radio icon unless necessary, don't create radios in vending machines until needed -- Saves 0.091s of init time (#3546)
* Don't update radio icon unless necessary, don't create radios in vending machines until needed -- Saves 0.091s of init time (#71682)

#71006 introduced a large regression in init times by forcing them to
update their icons for every single proc being called. This is very
slow. Defers the icon creation until initialize is finished, and only if
necessary.

Also learned that basically everything has a radio by making a station
bounced radio and putting it inside them. Yay...Removes the one in
vending machines until it's actually necessary. Was thinking of making
it a PDA message instead but nah

* Don't update radio icon unless necessary, don't create radios in vending machines until needed -- Saves 0.091s of init time

Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>