[MIRROR] [MDB IGNORE] fixes cleaning animation disappearing if you clean something that's already being cleaned (#3549)
* fixes cleaning animation disappearing if you clean something that's already being cleaned (#71673)

## About The Pull Request
1: Start cleaning something.
2: Clean the same thing again before the original cleaning manages to
finish. This removes the cleaning bubbles animation early, before the
actual cleaning is done. You can also spam the start cleaning message
with this.

The cleaning overlay is removed after the do_after that makes you wait
the time it takes to clean. When you're already waiting and you clean it
a second time the do_after returns instantly and removes the overlay
while the first one is still waiting.

But it seems that we don't even need to bother with any of that code if
we're already cleaning the thing, so we can just return early and
prevent the issue.
## Why It's Good For The Game
Fixes a graphical bug.
## Changelog
:cl:
fix: trying to clean something while it's already being cleaned will no
longer prematurely remove the cleaning animation
/:cl:

* fixes cleaning animation disappearing if you clean something that's already being cleaned

Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>