[MIRROR] [MDB IGNORE] captain can no longer steal cargo goodies (#3550)
* captain can no longer steal cargo goodies (#71649)

## About The Pull Request

Last night while messing around with some cargo order stuff I used the
captains ID while inside its PDA and saw an unfamiliar box.

![image](https://user-images.githubusercontent.com/116288367/205162064-7cd08de7-8dcc-4e47-b04b-492ecd113271.png)

Then I experimented and found out the following.
The Captains PDA can open cargo goodies.
The EMAG can also open cargo goodies.
When it does these things, the code goes through to the parent of the
order lockboxes and runs the /lockbox/attackby which means it goes
through access which means that it recognizes that the captain has
access and lets the captain open it.
This also updates their sprites into icon_unlocked and icon_broken which
don't exist for the secure briefcase that is used for cargo goodies.


![image](https://user-images.githubusercontent.com/116288367/205163658-3d88abe8-7620-428b-ace2-d1abb8b0ef96.png)
In this PR I added in a icon_broken because I think that it is fine to
emag goodies and that if so it should be consistent with the other
emag'd locked boxes.

I then made a small change that stops the captain from being able to
open goodies that do not belong to them, where it now reads the ID
inside of the PDA (or wallet) to compare bank accounts instead of
returning to the parent's version of the code.

Then I added screentips where if you have an ID (or something that
contains an ID) you can attempt to unlock a lockbox, it doesn't check if
you actually CAN unlock it because it's probably fine for screentips to
not be smart like that.

## Why It's Good For The Game

I think that the captain should not be able to steal cargo goodies and
it is against the spirit of how cargo goodies are coded. Emaging them is
fine, but it shouldn't turn into a completely different item if it is
emag'd.

Also personally I don't think that only being able to open cargo goodies
with the ID in hand is like, compelling enough to stay.

Screentips. 

## Changelog

:cl:
qol: lockboxes have screentips,
qol: cargo goodies can be opened with your ID even if it is inside a PDA
or Wallet
fix: captain can no longer open cargo goodies
imageadd: secure lockboxes have a broken sprite when emag'd
/:cl:

* captain can no longer steal cargo goodies

Co-authored-by: Sol N <116288367+flowercuco@users.noreply.github.com>