[MIRROR] [MDB IGNORE] Makes palladium synthate catalyst easier to create (#3585)
* Makes palladium synthate catalyst easier to create (#71797)

## About The Pull Request

As it stands, Palladium synthate catalyst is extremely hard to make
efficiently. The minimum temperature required for this reagent is set to
320K, but the issue is that Plasma, one of its ingredients, turns into
gas around 323-324K. Not only that, but the optimal temperature is set
to 600K, meaning even when its being made, it goes at a painfully slow
rate. As a result, there is an extremely slim temperature margin in
which this can be made without accidentally making plasma gas.

This PR serves to fix that issue by lowering the minimal temperature
threshold to 200K, and lowering the optimal temperature to 500K. While
you still have to pay attention while making it, at least the margin for
error is improved.

## Why It's Good For The Game

By making Palladium synthate catalyst easier to make, it expands more
options for chemists to utilize this reagent, such as in plumbing loops
or individual reactions.

## Changelog
:cl:
balance: Tweaks the minimal temperature threshold and optimal
temperature for Palladium synthate catalyst.
/:cl:

* Makes palladium synthate catalyst easier to create

Co-authored-by: MGOOOOOO <97645027+MGOOOOOO@users.noreply.github.com>