[MIRROR] [MDB IGNORE] Fixes many basic mobs not being vulnerable to extreme atmospheres and temperatures by making it the default (#3604)
* Fixes many basic mobs not being vulnerable to extreme atmospheres and temperatures by making it the default (#71817)

## About The Pull Request
**In short:**
This PR causes cows, pigs, rabbits, sheep, cockroaches, mothroaches and
mice to take damage in dangerous atmospheres and temperatures.

**In long:**
So simple mobs take damage from extreme atmospheres and temperatures by
default, this is not the case with basic mobs however. Basic mobs can be
given the `atmos_requirements` and `basic_body_temp_sensitive` elements
to accomplish this, but these are missing on a number of basic mobs that
probably shouldn't be able to survive in space.

This PR gives basic mobs the `atmos_requirements` and
`basic_body_temp_sensitive` elements by default, using the same values
as default simple mobs do. These values have been overridden in the
cases where this was done by the original simple mob (cockroaches and
lavaland mobs) or where the basic mob already had one of the elements
with different values (space bunnies and void puppies).

Not every mob is vulnerable to temperature or atmosphere though. Setting
`unsuitable_atmos_damage` to 0 will make it so the mob isn't given the
`atmos_requirements` element. Setting both
`minimum_survivable_temperature` and `maximum_survivable_temperature` to
0 will make it so the mob isn't given the `basic_body_temp_sensitive`
element.

I'm not sure if this is the best way of doing this. I don't know if mobs
dying to extreme atmos/temp by default was left behind with simple mobs
for a reason or if it was just forgotten about.
## Why It's Good For The Game
Cows, pigs, rabbits, sheep, cockroaches, mothroaches and mice are
probably not intended to be space-faring animals.

Basic mobs matching the same default atmosphere and temperature
vulnerabilities as simple mobs makes it harder to overlook that when
porting them to basic mobs.
## Changelog
:cl:
fix: cows, pigs, rabbits, sheep, cockroaches, mothroaches and mice now
take damage in hazardous atmospheres and temperatures
/:cl:

* Fixes many basic mobs not being vulnerable to extreme atmospheres and temperatures by making it the default

Co-authored-by: kawoppi <94711066+kawoppi@users.noreply.github.com>