[MIRROR] [MDB IGNORE] Refactor /ui_state/new_player_state to what it's designed for (#3715)
* Refactor /ui_state/new_player_state to what it's designed for (#71959)

## About The Pull Request

`/datum/ui_state/new_player_state` now no longer lets users with R_ADMIN
permission see the UI. Changed `/datum/interview` to return always_state
when an admin watches an interview because it is important there.
`/datum/latejoin_menu` now uses new_player_state instead of its own
solution.
## Why It's Good For The Game

`/datum/ui_state/new_player_state` is designed to let lobby players
interact with the UI, so it does not make sense for it to make a weird
exception for admins too. My theory is that this was an oversight when
this was first implemented, since only `/datum/interview` used this
ui_state. It doesn't and it shouldn't be like this though. Code clarity.
## Changelog
No playerfacing changes.

* Refactor /ui_state/new_player_state to what it's designed for

Co-authored-by: distributivgesetz <distributivgesetz93@gmail.com>