[MIRROR] [MDB IGNORE] Fixes inability to declare war as nukies (#3844)
* Fixes inability to declare war as nukies (#72217)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

War item checked for being on the CC zlevel which is no longer where the
nukie base is loaded.
Instead it now checks for being in a reserved area, which it will be if
loaded in __correctly__

fixes https://github.com/tgstation/tgstation/issues/72214

Also restricts loading lazy templates until the game has fully started,
i've noticed that trying to load pregame has a tendency to get stuck in
a runtime loop when SSicon_smoothing fires at the same time loading is
occurring

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

War Good

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: War can once again be declared
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* Fixes inability to declare war as nukies

Co-authored-by: Zephyr <12817816+ZephyrTFA@users.noreply.github.com>