[MIRROR] [MDB IGNORE] Fixes summon guns / magic from the secrets menu (#3916)
* Fixes summon guns / magic from the secrets menu (#72294)

## About The Pull Request

The proc takes a mob, not a client. `holder`'s a client. 

This is fine if it ends up being `null` for some reason

It's also completely viable to just not pass a mob and keep it `null`,
so that's option B

## Why It's Good For The Game

I could've sworn this worked when I tested it last but 🤷 

## Changelog

:cl: Melbert
fix: Secrets menu Summon Guns / Summon Magic work again.
/:cl:

Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>

* Fixes summon guns / magic from the secrets menu

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>