[MIRROR] [MDB IGNORE] Fix: DNA Infuser & Infusions (#4119)
* Fix: DNA Infuser & Infusions (#72688)

## About The Pull Request

This PR bug-fixes the DNA Infuser, the Goliath infusion, and fixes
examine-text grammar for the infusion organs.

Fixes #72350
Fixes #72738

Granular list of bugs and fixes:
- Fixed a bug which was causing the Goliath Infusion's "hammer" attack
bonus to stop working, which was caused by a minor developer oversight
in `/obj/item/goliath_infuser_hammer/attack` in which it sets the bonus
to 0. I also micro-optimized some of the bonus damage procs.
- Fixed incorrect grammar/pronoun usage in
`/datum/element/noticable_organ` and when adding the Element, as seen in
the Carp, Rat, and Goliath infusion organs.
- Fixed the DNA Infuser infusing into non-organic/DNA-less species by
adding conditionals to check for the mob biotype and validity of its
DNA.
- Fixed the DNA Infuser infusing into non-existent organs/brains. With
the changes, you must have lungs in order to get the carp-lungs, etc...
- Fixed the DNA Infuser not allowing dead creatures to be added. This
was due to a minor developer oversight inside
`/obj/machinery/dna_infuser/MouseDrop_T` wherein an extra conditional
was added.

## Why It's Good For The Game

There are some bugs in the Goliath infusion code which caused it to
deactivate its bonus attack damage, and this PR introduces a fix for
that. This PR also makes some grammar corrections to the rest of the
infusions, and adds an organic/DNA requirement to the DNA Infuser
machine.

## Changelog

:cl: A.C.M.O.
fix: Fixed the DNA Infuser's Goliath Infusion and its organ set bonus.
The Goliath hammer will now apply bonus attack damage as expected.
fix: Fixed incorrect grammar and pronoun usage in the DNA Infuser's
noticeable organs examine-text.
fix: Fixed the DNA Infuser giving mutated organs to androids and other
non-organic/DNA-less species.
fix: Fixed the DNA Infuser mutating organs which the occupant doesn't
have.
fix: Fixed the DNA Infuser not allowing addition of dead creatures.
/:cl:

* Fix: DNA Infuser & Infusions

Co-authored-by: Dani Glore <fantasticdragons@gmail.com>