Fix for plasmamen potentially spawning without internals, suits, etc (#72609)
## About The Pull Request

Currently there is a proc, 'give_important_for_life', which was unused
except in some downstreams, that was not working properly. It was meant
to be used for ensuring plasmamen get their gear upon spawning.

Fixed the issue with that proc and added it as a safety measure for
plasmamen who spawn without race-specific job outfits for whatever
reason (usually because one was not defined for that job). Also handles
plasmamen whose jobs are set to null for whatever reason (event spawners
perhaps?)

## Why It's Good For The Game

Fixes a proc that was bugged. Now plasmamen should get the generic
plasmamen outfit if their job does not specify a plasmamen version, thus
no more immediate combustion!

Fixes #60382 it would seem, tested and no issues were found.

## Changelog

:cl:
fix: fixed give_important_for_life proc in species.dm, which is supposed
to be used to help ensure plasmamen (and potentially other races) are
always spawned with internals and such
fix: fixed issue where plasmamen who are spawned without a
plasmaman-specific job outfit could spawn without internals and a suit,
and just start dying immediately
/:cl: