[MIRROR] [MDB IGNORE] Fiox firebots extinguishing themselves while also not on fire (fixes #71390) (#4179)
Original PR: https://github.com/tgstation/tgstation/pull/72707
-----
## About The Pull Request

Fixes #71390

Firebots returned themselves if they found themselves in the
`process_scan()` override. The override usually checks to see if the
found target is on fire, to then extinguish it. This leads to the
firebots attempting to constantly extinguish themselves.

I'm not sure why this behaviour was here, as the other bots don't seem
to return themselves in their `process_scan()` overrides. If someone has
any insights on why this should be here and I should fix it differently,
go ahead. The only alternative I see right now would be to return `null`
on finding that the current scan object is `src`, which would shorten
the program path a little.

## Why It's Good For The Game

Firebots extinguish people on fire instead of themselves.

## Changelog

:cl:
fix: Firebots no longer attempt to extinguish themselves constantly
/:cl:

Co-authored-by: Inari-Whitebear <inari_whitebear@fastmail.jp>