[MIRROR] [MDB IGNORE] Updates Asteroid Area Sprite (#4188)
Original PR: https://github.com/tgstation/tgstation/pull/72789
-----

## About The Pull Request

I didn't review it before I merged it (sorry) and it had a not great
outline, so I changed that and fixed the centering. While I was at it, I
changed the color scheme so it's not so garish to look at in a map
editor.

Used to look like this:


![image](https://user-images.githubusercontent.com/34697715/213097257-a6643b25-884b-4470-869e-42833711dca3.png)
## Why It's Good For The Game

Now looks like this:


![image](https://user-images.githubusercontent.com/34697715/213097315-5346f65e-5d86-4602-947b-63c76cd7b404.png)

Much calmer to the eyes rather than a vast sea of blue. Better for
editing maps.
## Changelog
Nothing that concerns players.

Co-authored-by: san7890 <the@san7890.com>