Add signal override back as a runtime (#72561)
We're not seeing this on CI and I'm not seeing it on my local DD logs
unless I go out of my way to check. Keeping separate file means it's
easy to go through, but this indicates a bug that needs to be fixed and
so it should still be in runtime.

Doesn't revert #70034, as it still keeps log_signal.