Package delivery sorter fix, autodocs (#72865)
## About The Pull Request

On some maps, there is a special disposal sorting pipe, which checks if
the `tomail` var of a disposal holder. However, this has been only be
set on the parent disposal object. The delivery chute and the disposal
bin overrode this behaviour without a parent call. I have elected to
remove the two identical overrides, and move their contents to the
parent, while also swapping out the `1` to the correct sorting define of
the same value. I have also updated the disposal holder object merging
proc.

I have also autodocced proc and var definiton in the files I have
touched.

## Why It's Good For The Game

Now once again, the wrapper sorter objects will detect wrapped packages,
and on maps that support them, will be redirected to the mail outlet
chute (along with any other item in the same holder object as usual),
instead of the trash outlet chute.


![image](https://user-images.githubusercontent.com/2676196/213875548-bf1e784f-be45-4cdf-af89-9a3fbc9bd6bd.png)

## Changelog

:cl:
fix: Bundles of disposal packages that contain a wrapped parcel will be
now successfully redirected by the wrapping sorters
/:cl: