[MIRROR] [MDB IGNORE] Allows MODSuit AIs to count towards Steal AI Traitor Objective (#4248)
Original PR: https://github.com/tgstation/tgstation/pull/72902
-----
## About The Pull Request

On the tin. It was simpler than it looked, just had to genericize the
behavior a tad bit (the proc is not called unless the type in either
targetitem or altitems is matched in
`/datum/objective/steal/check_completion()`). I decided that since the
current framework restricts us to only having a set range of items that
we may as well be safe and check the var that references the "stored"
AI, which works well enough.
## Why It's Good For The Game


![image](https://user-images.githubusercontent.com/34697715/213986912-506c8f27-131f-43e1-ab57-63635b9ed852.png)

Very cool!

Fixes #72205. Almost certainly an oversight, because if you create an AI
that is lawed to obey you regardless of anything in order to serve as
your modsuit buddy... it's just a pain and completely un-intuitive to
have to pop it out of the modsuit just so you can get the epic
greentext.
## Changelog
:cl:
fix: After over a year of continuous research, the Syndicate have
finally managed to crack Nanotrasen's Protocol for transferring AIs from
an intelliCard into a MODsuit, and vice-versa in order to extract the
valuable AI in a method fit-for-sale. In short: The "Steal a Functional
AI" objective will work if the "captured" AI is in a MODsuit.
/:cl:

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Jolly <70232195+Jolly-66@users.noreply.github.com>