[MIRROR] [MDB IGNORE] Add pickaxe to basic tools techweb (#4296)
Original PR: https://github.com/tgstation/tgstation/pull/72938
-----
## About The Pull Request
Adds the pickaxe to basic tools techweb that can be printed the cargo
lathe or autolathe.

## Why It's Good For The Game

This is a basic mining tool and should be available at roundstart.

It wasn't possible to create pickaxes, despite it being a T1 item. It
should be as easy to make as a shovel since these usually are mapped
together. I was tempted to put this into mining technology, but that
already has advanced drills that are T2 versions of the pickaxe.

## Changelog
:cl:
qol: Add pickaxe to basic tools techweb
/:cl:

---------

Co-authored-by: Tim <timothymtorres@gmail.com>