[MIRROR] [MDB IGNORE] Add advanced fire extinguisher to techweb (#4298)
Original PR: https://github.com/tgstation/tgstation/pull/72933
-----
## About The Pull Request
This adds the advanced fire extinguisher to the `Experimental Tools`
node.

## Why It's Good For The Game

There are several advanced engineering tools in this node. This is an
item that is considered a better version of a regular fire extinguisher
for atmos techs. It makes sense to put this with the other upgraded
tools in techweb, since ya know, it's technically an advanced version.
It does spawn empty, so it needs to be refilled with a foam tank
dispenser.

## Changelog
:cl:
qol: Add advanced fire extinguisher to techweb
/:cl:

---------

Co-authored-by: Tim <timothymtorres@gmail.com>